backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==3.8.0
django-environ==0.10.0
django-filter==2.4.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rest-framework-generic-relations==2.1.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
psycopg2-binary==2.9.7